Compute confetti sparkle position locally
Fix confetti is not behind text on ipad.

Instead of using initial positions, set a local position. This requires
more tuning (_Victory_AnimSpacing), but ensures that confetti is
positioned correctly regardless of screen size or changes to size (or
orientation).

Added a new group to make the base y off the bottom edge of the screen.
I think the positions end up the same as before.